@@ -28,78 +28,78 @@
     "new": "plop"
   },
   "dependencies": {
-    "@angular/animations": "~9.1.11",
-    "@angular/cdk": "^9.2.4",
-    "@angular/common": "~9.1.11",
-    "@angular/compiler": "~9.1.11",
-    "@angular/core": "~9.1.11",
-    "@angular/forms": "~9.1.11",
-    "@angular/platform-browser": "~9.1.11",
-    "@angular/platform-browser-dynamic": "~9.1.11",
-    "@angular/router": "~9.1.11",
-    "@aspnet/signalr": "^1.1.4",
+    "@angular/animations": "~11.2.5",
+    "@angular/cdk": "^11.2.6",
+    "@angular/common": "~11.2.5",
+    "@angular/compiler": "~11.2.5",
+    "@angular/core": "~11.2.5",
+    "@angular/forms": "~11.2.5",
+    "@angular/platform-browser": "~11.2.5",
+    "@angular/platform-browser-dynamic": "~11.2.5",
+    "@angular/router": "~11.2.5",
+    "@aspnet/signalr": "^1.0.26",
     "@ckeditor/ckeditor5-angular": "^2.0.1",
     "@ckeditor/ckeditor5-build-classic": "^27.0.0",
     "@ckeditor/ckeditor5-build-decoupled-document": "^27.0.0",
-    "@datorama/akita": "^5.2.1",
-    "@fortawesome/fontawesome-free": "^5.12.1",
+    "@datorama/akita": "^6.1.2",
+    "@fortawesome/fontawesome-free": "^5.15.3",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
-    "ajv": "^6.10.2",
-    "angular2-baidu-map": "^8.0.0",
-    "file-saver": "^2.0.2",
+    "ajv": "^7.2.3",
+    "angular2-baidu-map": "^8.1.1",
+    "file-saver": "^2.0.5",
     "modernizr": "^3.9.1",
     "moment": "^2.27.0",
     "ng-chat": "^3.0.3",
-    "ng-zorro-antd": "^9.3.0",
-    "ngx-bootstrap": "^5.6.1",
-    "ngx-toastr": "^13.0.0",
-    "oidc-client": "^1.10.1",
+    "ng-zorro-antd": "^11.3.0",
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-toastr": "^13.2.1",
+    "oidc-client": "^1.11.5",
     "qrious": "^4.0.2",
     "rxjs": "~6.5.5",
-    "screenfull": "^5.0.2",
-    "simple-line-icons": "^2.4.1",
+    "screenfull": "^5.1.0",
+    "simple-line-icons": "^2.5.5",
     "source-sans-pro": "^3.6.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "tslib": "^2.1.0",
+    "zone.js": "~0.11.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.7",
-    "@angular/cli": "~9.1.11",
-    "@angular/compiler-cli": "~9.1.11",
-    "@angular/language-service": "~9.1.11",
-    "@angularclass/hmr": "^2.1.3",
-    "@datorama/akita-ngdevtools": "^5.0.3",
-    "@types/node": "^12.11.1",
+    "@angular-devkit/build-angular": "~0.1102.6",
+    "@angular/cli": "~11.2.6",
+    "@angular/compiler-cli": "~11.2.5",
+    "@angular/language-service": "~11.2.5",
+    "@angularclass/hmr": "^3.0.0",
+    "@datorama/akita-ngdevtools": "^6.0.0",
+    "@types/node": "^14.14.35",
     "akita-schematics": "^2.0.1",
-    "codecov": "^3.6.1",
-    "codelyzer": "^5.1.2",
-    "gh-pages": "^2.1.1",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "codecov": "^3.8.1",
+    "codelyzer": "^6.0.1",
+    "gh-pages": "^3.1.0",
+    "jasmine-core": "~3.7.1",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "~6.3.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
     "less-bundle-promise": "^1.0.7",
-    "lint-staged": "^8.2.1",
-    "plop": "^2.7.3",
-    "prettier": "^1.18.2",
+    "lint-staged": "^10.5.4",
+    "plop": "^2.7.4",
+    "prettier": "^2.2.1",
     "prettier-stylelint": "^0.4.2",
-    "protractor": "~5.4.0",
-    "stylelint": "^11.1.1",
-    "stylelint-config-prettier": "^6.0.0",
+    "protractor": "~7.0.0",
+    "stylelint": "^13.12.0",
+    "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-rational-order": "^0.1.2",
-    "stylelint-config-standard": "^19.0.0",
-    "stylelint-declaration-block-no-ignored-properties": "^2.1.0",
-    "stylelint-order": "^3.1.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-declaration-block-no-ignored-properties": "^2.3.0",
+    "stylelint-order": "^4.1.0",
     "swagger-angular-codegen": "^0.1.8",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
+    "ts-node": "~9.1.1",
+    "tslint": "~6.1.3",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "~3.7.5",
-    "webpack-bundle-analyzer": "^3.6.0"
+    "typescript": "~4.0.3",
+    "webpack-bundle-analyzer": "^4.4.0"
   },
   "lint-staged": {
     "linters": {
