@@ -1,5 +1,5 @@
-import { Component, OnInit } from '@angular/core';
-import { NzModalService, NzMessageService } from 'ng-zorro-antd';
+import { Component, OnInit } from '@angular/core';import {  NzModalService } from 'ng-zorro-antd/modal';
+import { NzMessageService } from 'ng-zorro-antd/message';
 import { ActivatedRoute } from '@angular/router';
 import { PartnerProxyService } from 'src/api/appService';
 
