@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { NzModalService, NzMessageService } from 'ng-zorro-antd';
+import {  NzModalService } from 'ng-zorro-antd/modal';
+import { NzMessageService } from 'ng-zorro-antd/message';
+
 import { ActivatedRoute } from '@angular/router';
 import { SwiperProxyService } from 'src/api/appService';
 import { SwiperEditComponent } from './swiper-edit.component';
