@@ -14,7 +14,7 @@ export function getClientSettings(): UserManagerSettings {
         redirect_uri: window.location.origin + environment.SERVER_URL + '/callback.html',
         //post_logout_redirect_uri: window.location.origin + environment.SERVER_URL + '/signout-callback.html',
         response_type: 'id_token token',
-        scope: "address email openid phone profile role",
+        scope: "address email openid phone profile role SoMall",
         accessTokenExpiringNotificationTime: 20,
         automaticSilentRenew: false,
         filterProtocolClaims: true,
