@@ -1,5 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -15,18 +19,18 @@ namespace TT.SoMall.Data
         public ILogger<SoMallDbMigrationService> Logger { get; set; }
 
         private readonly IDataSeeder _dataSeeder;
-        private readonly ISoMallDbSchemaMigrator _dbSchemaMigrator;
+        private readonly IEnumerable<ISoMallDbSchemaMigrator> _dbSchemaMigrators;
         private readonly ITenantRepository _tenantRepository;
         private readonly ICurrentTenant _currentTenant;
 
         public SoMallDbMigrationService(
             IDataSeeder dataSeeder,
-            ISoMallDbSchemaMigrator dbSchemaMigrator,
+            IEnumerable<ISoMallDbSchemaMigrator> dbSchemaMigrators,
             ITenantRepository tenantRepository,
             ICurrentTenant currentTenant)
         {
             _dataSeeder = dataSeeder;
-            _dbSchemaMigrator = dbSchemaMigrator;
+            _dbSchemaMigrators = dbSchemaMigrators;
             _tenantRepository = tenantRepository;
             _currentTenant = currentTenant;
 
@@ -35,45 +39,151 @@ namespace TT.SoMall.Data
 
         public async Task MigrateAsync()
         {
+            try
+            {
+                if (DbMigrationsProjectExists() && !MigrationsFolderExists())
+                {
+                    AddInitialMigration();
+                    return;
+                }
+            }
+            catch (Exception e)
+            {
+                Logger.LogWarning("Couldn't determinate if any migrations exist : " + e.Message);
+            }
+
             Logger.LogInformation("Started database migrations...");
 
-            await MigrateHostDatabaseAsync();
+            await MigrateDatabaseSchemaAsync();
+            await SeedDataAsync();
 
-            var i = 0;
-            var tenants = await _tenantRepository.GetListAsync();
+            Logger.LogInformation($"Successfully completed host database migrations.");
+
+            var tenants = await _tenantRepository.GetListAsync(includeDetails: true);
+
+            var migratedDatabaseSchemas = new HashSet<string>();
             foreach (var tenant in tenants)
             {
-                i++;
-
                 using (_currentTenant.Change(tenant.Id))
                 {
-                    Logger.LogInformation($"Migrating {tenant.Name} database schema... ({i} of {tenants.Count})");
-                    await MigrateTenantDatabasesAsync(tenant);
-                    Logger.LogInformation($"Successfully completed {tenant.Name} database migrations.");
+                    if (tenant.ConnectionStrings.Any())
+                    {
+                        var tenantConnectionStrings = tenant.ConnectionStrings
+                            .Select(x => x.Value)
+                            .ToList();
+
+                        if (!migratedDatabaseSchemas.IsSupersetOf(tenantConnectionStrings))
+                        {
+                            await MigrateDatabaseSchemaAsync(tenant);
+
+                            migratedDatabaseSchemas.AddIfNotContains(tenantConnectionStrings);
+                        }
+                    }
+
+                    await SeedDataAsync(tenant);
+                }
+
+                Logger.LogInformation($"Successfully completed {tenant.Name} tenant database migrations.");
+            }
+
+            Logger.LogInformation("Successfully completed all database migrations.");
+            Logger.LogInformation("You can safely end this process...");
+        }
+
+        private async Task MigrateDatabaseSchemaAsync(Tenant tenant = null)
+        {
+            Logger.LogInformation(
+                $"Migrating schema for {(tenant == null ? "host" : tenant.Name + " tenant")} database...");
+
+            foreach (var migrator in _dbSchemaMigrators)
+            {
+                await migrator.MigrateAsync();
+            }
+        }
+
+        private async Task SeedDataAsync(Tenant tenant = null)
+        {
+            Logger.LogInformation($"Executing {(tenant == null ? "host" : tenant.Name + " tenant")} database seed...");
+
+            await _dataSeeder.SeedAsync(tenant?.Id);
+        }
+
+        private bool DbMigrationsProjectExists()
+        {
+            var dbMigrationsProjectFolder = GetDbMigrationsProjectFolderPath();
+
+            return dbMigrationsProjectFolder != null;
+        }
+
+        private bool MigrationsFolderExists()
+        {
+            var dbMigrationsProjectFolder = GetDbMigrationsProjectFolderPath();
+
+            return Directory.Exists(Path.Combine(dbMigrationsProjectFolder, "migrations"));
+        }
+
+        private void AddInitialMigration()
+        {
+            Logger.LogInformation("Creating initial migration...");
+
+            string argumentPrefix;
+            string fileName;
+
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) || RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            {
+                argumentPrefix = "-c";
+                fileName = "/bin/bash";
+            }
+            else
+            {
+                argumentPrefix = "/C";
+                fileName = "cmd.exe";
+            }
+
+            var procStartInfo = new ProcessStartInfo( fileName,
+                $"{argumentPrefix} \"abp create-migration-and-run-migrator \"{GetDbMigrationsProjectFolderPath()}\"\""
+            );
+
+            try
+            {
+                Process.Start(procStartInfo);
+            }
+            catch (Exception)
+            {
+                throw new Exception("Couldn't run ABP CLI...");
+            }
+        }
+
+        private string GetDbMigrationsProjectFolderPath()
+        {
+            var slnDirectoryPath = GetSolutionDirectoryPath();
+
+            if (slnDirectoryPath == null)
+            {
+                throw new Exception("Solution folder not found!");
+            }
+
+            var srcDirectoryPath = Path.Combine(slnDirectoryPath, "src");
+
+            return Directory.GetDirectories(srcDirectoryPath)
+                .FirstOrDefault(d => d.EndsWith(".DbMigrations"));
+        }
+
+        private string GetSolutionDirectoryPath()
+        {
+            var currentDirectory = new DirectoryInfo(Directory.GetCurrentDirectory());
+
+            while (Directory.GetParent(currentDirectory.FullName) != null)
+            {
+                currentDirectory = Directory.GetParent(currentDirectory.FullName);
+
+                if (Directory.GetFiles(currentDirectory.FullName).FirstOrDefault(f => f.EndsWith(".sln")) != null)
+                {
+                    return currentDirectory.FullName;
                 }
             }
 
-            Logger.LogInformation("Successfully completed database migrations.");
-        }
-
-        private async Task MigrateHostDatabaseAsync()
-        {
-            Logger.LogInformation("Migrating host database schema...");
-            await _dbSchemaMigrator.MigrateAsync();
-
-            Logger.LogInformation("Executing host database seed...");
-            await _dataSeeder.SeedAsync();
-
-            Logger.LogInformation("Successfully completed host database migrations.");
-        }
-
-        private async Task MigrateTenantDatabasesAsync(Tenant tenant)
-        {
-            Logger.LogInformation($"Migrating schema for {tenant.Name} database...");
-            await _dbSchemaMigrator.MigrateAsync();
-
-            Logger.LogInformation($"Executing {tenant.Name} tenant database seed...");
-            await _dataSeeder.SeedAsync(tenant.Id);
+            return null;
         }
     }
 }
