@@ -7,6 +7,7 @@ using TT.SoMall.Users;
 using Volo.Abp.Data;
 using Volo.Abp.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore.Modeling;
+using Volo.Abp.Identity;
 using Volo.Abp.Users.EntityFrameworkCore;
 
 namespace TT.SoMall.EntityFrameworkCore
@@ -44,7 +45,8 @@ namespace TT.SoMall.EntityFrameworkCore
 
             builder.Entity<AppUser>(b =>
             {
-                b.ToTable("AbpUsers"); //Sharing the same table "AbpUsers" with the IdentityUser
+                b.ToTable(AbpIdentityDbProperties.DbTablePrefix + "Users"); //Sharing the same table "AbpUsers" with the IdentityUser
+                
                 b.ConfigureByConvention();
                 b.ConfigureAbpUser();
                 //Moved customization to a method so we can share it with the SoMallMigrationsDbContext class
